print the first 98 fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
  */
 int main(void)
 {
-	float prev, curr, temp;
+	double prev, curr, temp;
 	int i;
 
 	i = 1;
